+++ bika/lims/content/analysisrequest.py
@@ -130,6 +130,7 @@ from bika.lims.workflow import getTransitionUsers
 
 IMG_SRC_RX = re.compile(r'<img.*?src="(.*?)"')
 IMG_DATA_SRC_RX = re.compile(r'<img.*?src="(data:image/.*?;base64,)(.*?)"')
+FINAL_STATES = ["published", "retracted", "rejected", "cancelled"]
 
 
 # SCHEMA DEFINITION
@@ -248,13 +249,6 @@ schema = BikaSchema.copy() + Schema((
                         "sort_on": "sortable_title",
                         "sort_order": "ascending"},
             showOn=True,
-            add_button={
-                    'visible': True,
-                    'url': 'clients/createObject?type_name=Client',
-                    'return_fields': ['Title'],
-                    'js_controllers': ['#client-base-edit'],
-                    'overlay_handler': 'ClientOverlayHandler',
-                }
         ),
     ),
 
@@ -2437,6 +2431,35 @@ class AnalysisRequest(BaseFolder, ClientAwareMixin):
             html = html.replace(src, urljoin(base_url, src))
 
         return html
+
+    def getProgress(self):
+        """Returns the progress in percent of all analyses
+        """
+        review_state = api.get_review_status(self)
+
+        # Consider final states as 100%
+        # https://github.com/senaite/senaite.core/pull/1544#discussion_r379821841
+        if review_state in FINAL_STATES:
+            return 100
+
+        numbers = self.getAnalysesNum()
+
+        num_analyses = numbers[1] or 0
+        if not num_analyses:
+            return 0
+
+        # [verified, total, not_submitted, to_be_verified]
+        num_to_be_verified = numbers[3] or 0
+        num_verified = numbers[0] or 0
+
+        # 2 steps per analysis (submit, verify) plus one step for publish
+        max_num_steps = (num_analyses * 2) + 1
+        num_steps = num_to_be_verified + (num_verified * 2)
+        if not num_steps:
+            return 0
+        if num_steps > max_num_steps:
+            return 100
+        return (num_steps * 100) / max_num_steps
 
 
 registerType(AnalysisRequest, PROJECTNAME)